email/pyproject.toml
[project]
name = "phishguard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dnspython>=2.8.0",
    "fastapi>=0.136.1",
    "google-api-python-client>=2.194.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "joblib>=1.5.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "supabase>=2.29.0",
    "uvicorn[standard]>=0.46.0",
]
